various fixes + move channel_list to primitives::sentry
