+++ primitives/src/sentry.rs
@@ -215,6 +215,40 @@ impl fmt::Display for ChannelReport {
     }
 }
 
+pub mod channel_list {
+    use crate::ValidatorId;
+    use chrono::serde::ts_seconds::deserialize as ts_seconds;
+    use chrono::{DateTime, Utc};
+    use serde::Deserialize;
+
+    #[derive(Debug, Deserialize)]
+    pub struct ChannelListQuery {
+        #[serde(default = "default_page")]
+        pub page: u64,
+        /// filters the list on `valid_until >= valid_until_ge`
+        /// It should be the same timestamp format as the `Channel.valid_until`: **seconds**
+        #[serde(
+            deserialize_with = "ts_seconds",
+            default = "Utc::now",
+            rename = "validUntil"
+        )]
+        pub valid_until_ge: DateTime<Utc>,
+        pub creator: Option<String>,
+        /// filters the channels containing a specific validator if provided
+        pub validator: Option<ValidatorId>,
+    }
+
+    #[derive(Debug, Deserialize)]
+    #[serde(rename_all = "camelCase")]
+    pub struct LastApprovedQuery {
+        pub with_heartbeat: Option<String>,
+    }
+
+    fn default_page() -> u64 {
+        0
+    }
+}
+
 #[cfg(feature = "postgres")]
 mod postgres {
     use super::{
